requirements.txt
Flask
Flask-SQLAlchemy
beautifulsoup4
requests
lxml
python-dotenv
selenium
webdriver-manager
PyMySQL
undetected-chromedriver